% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instrument_table.R
\name{get_instrument_table}
\alias{get_instrument_table}
\title{Get Instrument Table}
\usage{
get_instrument_table(keyName = "cozi_instruments_id", force = F)
}
\arguments{
\item{keyName}{string default cozi_instruments_id. lets the user override the name of their stored key}

\item{force}{boolean default false. When true re-download the instrument table even if it is cached in the package environment}
}
\description{
Returns a tibble containing name, ip, port and serial for all the instruments
support by the logger
Right now this downloads the cozi_instruments spreadsheet from google drive.
The user needs have "cozi_instruments_id" as the sheets_id in {{keyring}} such that
\code{keyring::key_get("cozi_instruments_id")} returns the sheet id.
This will certainly change when the instrument data finds a better home
}
\author{
W. S. Drysdale
}
