% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_scale_defaults.R
\name{graph_scale_defaults}
\alias{graph_scale_defaults}
\title{Graph Scale Defaults}
\usage{
graph_scale_defaults()
}
\description{
Returns a table of x and y scales that can be used as global variables to control graphing
}
\author{
W. S. Drysdale
}
