% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_poll.R
\name{make_poll}
\alias{make_poll}
\title{Make poll}
\usage{
make_poll(instrumentNames)
}
\arguments{
\item{instrumentNames}{vector of instrument names matching those in the table defined by \code{get_instrument_table()}}
}
\description{
Creates the daq sequence(s) "Poll_Instrument". \cr
Returns a list of strings as each needs to be written as a separate sequence
}
\author{
W. S. Drysdale
}
