% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pollAll.R
\name{make_pollAll}
\alias{make_pollAll}
\title{Make pollAll}
\usage{
make_pollAll(instrumentNames)
}
\arguments{
\item{instrumentNames}{vector of instrument names matching those in the table defined by \code{get_instrument_table()}}
}
\description{
Creates the daq sequence(s) "Poll_Instrument". \cr
This sequence triggers the various poll_instrument sequences every 10 seconds
}
\author{
W. S. Drysdale
}
