% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sequences.R
\name{make_sequences}
\alias{make_sequences}
\title{Make Sequences}
\usage{
make_sequences(
  instrumentNames,
  fpRoot,
  filePrefix,
  historyLength = 21600,
  outputPath
)
}
\arguments{
\item{instrumentNames}{vector of instrument names matching those in the table defined by \code{get_instrument_table()}}

\item{fpRoot}{path to the folder in which to save data}

\item{filePrefix}{name to prefix to file ahead of the start date.}

\item{historyLength}{numeric default 21600 - DAQFactory \code{channelName.HistoryLength} - how long (in seconds) should the channel keep in memory}

\item{outputPath}{path to folder where sequeces should be saved}
}
\description{
Creates init, logging, pollAll and various poll_instrument seqences and saves as .txt files
}
\author{
W. S. Drysdale
}
